apps/ml-services/requirements.txt
# Core ML Libraries from MDK - Updated for Python 3.12 compatibility
pandas>=2.1.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
torch>=2.0.0
xgboost>=2.0.0
prophet>=1.1.0
joblib>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# For Temporal and Modal
temporalio
modal
boto3
fastapi
requests

# For Database Connection (from Python)
psycopg2-binary

# Updated dependencies for Python 3.12 compatibility
setuptools>=68.0.0
Cython>=3.0.0
numpy>=1.24.0
